% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file_prefix.R
\name{add_file_prefix}
\alias{add_file_prefix}
\title{add_file_prefix
add a prefix to filename}
\usage{
add_file_prefix(
  file_path,
  file_name = NULL,
  output_dir = NULL,
  width = 2,
  date = FALSE
)
}
\description{
add_file_prefix
add a prefix to filename
}
