% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPartSim.R
\name{calcPartSim}
\alias{calcPartSim}
\title{partition_similarity}
\usage{
calcPartSim(p1, p2, ignore = 0, normalized = TRUE)
}
\arguments{
\item{p1}{- partition 1 - named vector}

\item{p2}{- partition 2 - named vector}
}
\value{
none
}
\description{
calculate pairwise module assignment similarity statistic
}
\examples{
partition_similarity(p1, p2)
}
\references{
Choobdar et al.,2019{https://www.ncbi.nlm.nih.gov/pubmed/31471613}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
