% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_modules.R
\name{check_modules}
\alias{check_modules}
\title{check_modules}
\usage{
check_modules(x, strength = "strong", stats = c(1:7), alpha = 0.05)
}
\arguments{
\item{selfPreservation}{result returned by NetRep.}
}
\value{
none
}
\description{
Check modules for evidence of preservation.
}
\examples{
check_modules(selfPreservation)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
