% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCytoscapeGraph.R
\name{createCytoscapeGraph}
\alias{createCytoscapeGraph}
\title{createCytoscapeGraph}
\usage{
createCytoscapeGraph(
  netw_g,
  ppi_g,
  nodes,
  module_name,
  n_cutoffs = 5000,
  netwdir = getwd(),
  imgsdir = getwd(),
  netw_layout = NULL
)
}
\description{
createCytoscapeGraph
}
