% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{ggtheme}
\alias{ggtheme}
\title{ggtheme}
\usage{
ggtheme()
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Set a theme for ggplots.
}
\examples{
ggtheme()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
