% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{gse}
\alias{gse}
\title{gse}
\usage{
gse(gene_list, GeneSetcollection, background = "given", ...)
}
\arguments{
\item{gene_list}{- list of entrez genes. elements of list coorespond to
different groups.}
}
\value{
none
}
\description{
A function to perform gense set enrichment analysis given a
}
\examples{
gse()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
