% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeKNNpep.R
\name{imputeKNNpep}
\alias{imputeKNNpep}
\title{imputeKNNpep}
\usage{
imputeKNNpep(
  tp,
  groupBy = NULL,
  samples_to_ignore = NULL,
  colID = "Abundance",
  max_percent_missing = 0.5,
  quiet = TRUE
)
}
\description{
imputeKNNpep
}
