% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{js}
\alias{js}
\title{js}
\usage{
js(x, y)
}
\arguments{
\item{x}{- first vector of items.}

\item{y}{- second vector of items.}
}
\value{
none
}
\description{
jacaard similarity of two vectors
}
\examples{
js(x, y)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
