% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleGOenrichment.R
\name{moduleGOenrichment}
\alias{moduleGOenrichment}
\title{moduleGOenrichment}
\usage{
moduleGOenrichment(
  partition,
  gene_map,
  GOcollection,
  useBackground = NULL,
  p.adjust = "FDR",
  partition.ids = "id",
  alpha = 0.05,
  ...
)
}
\value{
none
}
\description{
A function to perform GO enrichmnet analysis of protein co-expression
modules.
}
\examples{
moduleGOenrichment()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
