% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{plot_profile}
\usage{
plot_profile(
  protein,
  wt_color = "#47b2a4",
  mut_color = col2hex(c(R = 148, G = 33, B = 146)),
  scale_abundance = FALSE
)
}
\description{
plot_profile
}
