% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_index.R
\name{reset_index}
\alias{reset_index}
\title{reset_index}
\usage{
reset_index(partition)
}
\arguments{
\item{partition}{- a named vector describing the partition of a network.}
}
\value{
none
}
\description{
reset partition indices
}
\examples{
reset_index(partition)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
