% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silence.R
\name{silence}
\alias{silence}
\title{silence}
\usage{
silence(x)
}
\arguments{
\item{x}{- expression to be silenced.}
}
\value{
none
}
\description{
Function for supressing printed messages from a function.
}
\examples{
silence(myfun(x))
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
