% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write_excel}
\alias{write_excel}
\title{write_excel}
\usage{
write_excel(data, file, rowNames = FALSE, colNames = TRUE, ...)
}
\arguments{
\item{data}{(list, matrix, dataframe)}
}
\value{
None
}
\description{
Utility function to write data to excel. Data can be provided as a named list.
}
\examples{
write_excel()
}
\references{
\url{}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{save}
\keyword{workbook}
\keyword{write}
\keyword{write.excel}
\keyword{xls}
\keyword{xlsx}
